src/requirements.txt
django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==25.1.0
openai==2.29.0
psycopg2-binary==2.9.11
pyTelegramBotAPI==4.32.0
python-dotenv==1.2.1
python-logging-loki==0.3.1
redis==7.4.0
yookassa==3.10.0
